requirements.txt
discord.py==2.5.2
pandas==2.3.1
PyMySQL==1.1.1
PyMySQL[rsa]
